unify matrix writing function
