+++ source/module_io/CMakeLists.txt
@@ -53,7 +53,6 @@ if(ENABLE_LCAO)
       read_dm.cpp
       read_wfc_nao.cpp
       write_wfc_nao.cpp
-      write_HS.cpp
       write_dm.cpp
       dos_nao.cpp
       output_dm.cpp
