+++ source/module_io/write_HS.h
@@ -10,65 +10,34 @@
 // mohan add this file 2010-09-10
 namespace ModuleIO
 {
-void saving_HS(const int istep,
-               const double* Hloc,
-               const double* Sloc,
-               const bool bit,
-               const int& out_hs,
-               const std::string& file_name,
-               const Parallel_Orbitals& pv,
-               bool tri);
-// tri : if set 1 , output triangle matrix ; if set 0 , output complete matrix .
+    /// @brief save a square matrix
+    /// @param[in] istep : the step of the calculation
+    /// @param[in] mat : the local matrix
+    /// @param[in] bit : true for binary, false for decimal
+    /// @param[in] tri : true for upper triangle, false for full matrix
+    /// @param[in] app : true for append, false for overwrite
+    /// @param[in] label : the symbol of the matrix, like "H", "S"
+    /// @param[in] file_name : the name of the output file
+    /// @param[in] pv : the 2d-block parallelization information
+    /// @param[in] drank : the rank of the current process
+    template<typename T>
+    void save_mat(const int istep,
+        const T* mat,
+        const int dim,
+        const bool bit,
+        const bool tri,
+        const bool app,
+        const std::string label,
+        const std::string& file_name,
+        const Parallel_2D& pv,
+        const int drank);
 
-void save_HS_triangle(const int istep,
-                      const double* H,
-                      const double* S,
-                      const bool bit,
-                      const std::string& file_name,
-                      const Parallel_Orbitals& pv);
-void save_HS_complete(const int istep,
-                      const double* H,
-                      const double* S,
-                      const bool bit,
-                      const std::string& file_name,
-                      const Parallel_Orbitals& pv);
-
-void save_HS_complex(const int istep,
-                     const std::complex<double>* H,
-                     const std::complex<double>* S,
-                     const bool bit,
-                     const std::string& file__name,
-                     const Parallel_Orbitals& pv,
-                     bool tri);
-
-void save_HSR_tr(const int current_spin, LCAO_Matrix& lm); // LiuXh add 2019-07-15
+    // comment out this function for not used
+    // void save_HSR_tr(const int current_spin, LCAO_Matrix& lm); // LiuXh add 2019-07-15
 
 // mohan comment out 2021-02-10
 // void save_HS_ccf(const int &iter, const int &Hnnz, const int *colptr_H, const int *rowind_H, 
 // const double *nzval_H, const double *nzval_S, bool bit);
-
-void saving_HS(const int istep,
-               std::complex<double>* Hloc,
-               std::complex<double>* Sloc,
-               bool bit,
-               const int& out_hs,
-               const std::string& file__name,
-               const Parallel_Orbitals& pv,
-               bool tri); // LiuXh, 2017-03-21
-
-void save_HS_complex_triangle(const int istep,
-                              std::complex<double>* H,
-                              std::complex<double>* S,
-                              const bool bit,
-                              const std::string& file__name,
-                              const Parallel_Orbitals& pv); // LiuXh, 2017-03-21
-
-void save_HS_complex_complete(const int istep,
-                              std::complex<double>* H,
-                              std::complex<double>* S,
-                              const bool bit,
-                              const std::string& file__name,
-                              const Parallel_Orbitals& pv); // LiuXh, 2017-03-21
 }
-
+#include "write_HS.hpp"
 #endif
